Added retry support for 500 error message
check_retry now checks if the entire URL returns an error message of 500. If so, we will try to visit the URL one more time in the future.

scrape function is modified to return the links set and a boolean signifying a potential retry. True: tell worker to retry. False: continue

In the worker, if retry is returned as True from scraper, attempt a retry.

In frontier, use a wrapper to manage adding urls the normal way vs forcing a link to be added. Links being marked complete are handled in the worker still.

Config add a file I made that forces a 500 internal service error.