Fixed fix_frontier to also check token hash database. Also, show a preview of links to delete by default.
fix_frontier now requires an explicit flag for the frontier and tokenhash dbs to be edited with the new is_valid function. Also, make sure the links are deleted in both the frontier and tokenhash dbs to ensure the links can be downloaded again if is_valid had an error prior.